+++ pkgs/objective_c/CHANGELOG.md
@@ -11,6 +11,8 @@
 - Fix a [bug](https://github.com/dart-lang/native/issues/3290) where missing
   debug symbols caused app store valiadtion warnings.
 - Removed some test-only utilities from the release dylib (fixes [#2999](https://github.com/dart-lang/native/issues/2999)).
+- Added `bool.toNSNumber()` extension method, `NSNumber.isBool` check and
+  converter support for bool.
 
 ## 9.3.0
 - `autoReleasePool` now returns the value produced by its callback.
@@ -18,7 +20,7 @@
 ## 9.2.5
 - Fix a [bug](https://github.com/dart-lang/native/issues/3011) by adding
   minimum OS version flags to the build script.
- 
+
 ## 9.2.4
 
 - Fix a [bug](https://github.com/dart-lang/native/issues/2990) build hook path